BGDIINF_SB-1857: Split search e2e tests in two
